refactor(code-examples): move all code examples related stuff into transformer package
