+++ packages/gatsby-theme-docs/gatsby-config.js
@@ -98,6 +98,14 @@ module.exports = (themeOptions = {}) => {
           path: path.resolve(`./src/releases`),
         },
       },
+      // Code examples
+      {
+        resolve: 'gatsby-source-filesystem',
+        options: {
+          name: 'codeExamples',
+          path: path.resolve(`./src/code-examples`),
+        },
+      },
 
       /**
        * Transformers for making content available in graphql queries
