+++ packages/gatsby-transformer-code-examples/package.json
@@ -14,5 +14,20 @@
     "gatsby-transformer",
     "code",
     "examples"
-  ]
+  ],
+  "main": "src/index.js",
+  "dependencies": {
+    "@commercetools-docs/ui-kit": "2.5.0",
+    "prop-types": "15.7.2"
+  },
+  "devDependencies": {
+    "gatsby": "2.21.11",
+    "react": "16.13.1",
+    "react-dom": "16.13.1"
+  },
+  "peerDependencies": {
+    "gatsby": "2.x",
+    "react": "16.x",
+    "react-dom": "16.x"
+  }
 }
