+++ packages/gatsby-transformer-code-examples/src/index.js
@@ -1,0 +1,2 @@
+export { default as CodeExample } from './code-example';
+export { default as MultiCodeExample } from './multi-code-example';
